+++ cloudauth-intl-20220809/alibabacloud_cloudauth_intl20220809/models.py
@@ -2540,9 +2540,13 @@ class InitializeRequest(TeaModel):
         callback_token: str = None,
         callback_url: str = None,
         crop: str = None,
+        date_of_birth: str = None,
+        date_of_expiry: str = None,
+        doc_page_config: List[str] = None,
         doc_scan_mode: str = None,
         doc_type: str = None,
         doc_video: str = None,
+        document_number: str = None,
         experience_code: str = None,
         face_picture_base_64: str = None,
         face_picture_url: str = None,
@@ -2550,6 +2554,7 @@ class InitializeRequest(TeaModel):
         id_spoof: str = None,
         id_threshold: str = None,
         language_config: str = None,
+        mrtdinput: str = None,
         merchant_biz_id: str = None,
         merchant_user_id: str = None,
         meta_info: str = None,
@@ -2565,15 +2570,20 @@ class InitializeRequest(TeaModel):
         show_guide_page: str = None,
         show_ocr_result: str = None,
         style_config: str = None,
+        use_nfc: str = None,
     ):
         self.app_quality_check = app_quality_check
         self.authorize = authorize
         self.callback_token = callback_token
         self.callback_url = callback_url
         self.crop = crop
+        self.date_of_birth = date_of_birth
+        self.date_of_expiry = date_of_expiry
+        self.doc_page_config = doc_page_config
         self.doc_scan_mode = doc_scan_mode
         self.doc_type = doc_type
         self.doc_video = doc_video
+        self.document_number = document_number
         self.experience_code = experience_code
         self.face_picture_base_64 = face_picture_base_64
         self.face_picture_url = face_picture_url
@@ -2581,6 +2591,7 @@ class InitializeRequest(TeaModel):
         self.id_spoof = id_spoof
         self.id_threshold = id_threshold
         self.language_config = language_config
+        self.mrtdinput = mrtdinput
         self.merchant_biz_id = merchant_biz_id
         self.merchant_user_id = merchant_user_id
         self.meta_info = meta_info
@@ -2597,6 +2608,7 @@ class InitializeRequest(TeaModel):
         self.show_guide_page = show_guide_page
         self.show_ocr_result = show_ocr_result
         self.style_config = style_config
+        self.use_nfc = use_nfc
 
     def validate(self):
         pass
@@ -2617,12 +2629,20 @@ class InitializeRequest(TeaModel):
             result['CallbackUrl'] = self.callback_url
         if self.crop is not None:
             result['Crop'] = self.crop
+        if self.date_of_birth is not None:
+            result['DateOfBirth'] = self.date_of_birth
+        if self.date_of_expiry is not None:
+            result['DateOfExpiry'] = self.date_of_expiry
+        if self.doc_page_config is not None:
+            result['DocPageConfig'] = self.doc_page_config
         if self.doc_scan_mode is not None:
             result['DocScanMode'] = self.doc_scan_mode
         if self.doc_type is not None:
             result['DocType'] = self.doc_type
         if self.doc_video is not None:
             result['DocVideo'] = self.doc_video
+        if self.document_number is not None:
+            result['DocumentNumber'] = self.document_number
         if self.experience_code is not None:
             result['ExperienceCode'] = self.experience_code
         if self.face_picture_base_64 is not None:
@@ -2637,6 +2657,8 @@ class InitializeRequest(TeaModel):
             result['IdThreshold'] = self.id_threshold
         if self.language_config is not None:
             result['LanguageConfig'] = self.language_config
+        if self.mrtdinput is not None:
+            result['MRTDInput'] = self.mrtdinput
         if self.merchant_biz_id is not None:
             result['MerchantBizId'] = self.merchant_biz_id
         if self.merchant_user_id is not None:
@@ -2667,6 +2689,8 @@ class InitializeRequest(TeaModel):
             result['ShowOcrResult'] = self.show_ocr_result
         if self.style_config is not None:
             result['StyleConfig'] = self.style_config
+        if self.use_nfc is not None:
+            result['UseNFC'] = self.use_nfc
         return result
 
     def from_map(self, m: dict = None):
@@ -2681,12 +2705,20 @@ class InitializeRequest(TeaModel):
             self.callback_url = m.get('CallbackUrl')
         if m.get('Crop') is not None:
             self.crop = m.get('Crop')
+        if m.get('DateOfBirth') is not None:
+            self.date_of_birth = m.get('DateOfBirth')
+        if m.get('DateOfExpiry') is not None:
+            self.date_of_expiry = m.get('DateOfExpiry')
+        if m.get('DocPageConfig') is not None:
+            self.doc_page_config = m.get('DocPageConfig')
         if m.get('DocScanMode') is not None:
             self.doc_scan_mode = m.get('DocScanMode')
         if m.get('DocType') is not None:
             self.doc_type = m.get('DocType')
         if m.get('DocVideo') is not None:
             self.doc_video = m.get('DocVideo')
+        if m.get('DocumentNumber') is not None:
+            self.document_number = m.get('DocumentNumber')
         if m.get('ExperienceCode') is not None:
             self.experience_code = m.get('ExperienceCode')
         if m.get('FacePictureBase64') is not None:
@@ -2701,6 +2733,8 @@ class InitializeRequest(TeaModel):
             self.id_threshold = m.get('IdThreshold')
         if m.get('LanguageConfig') is not None:
             self.language_config = m.get('LanguageConfig')
+        if m.get('MRTDInput') is not None:
+            self.mrtdinput = m.get('MRTDInput')
         if m.get('MerchantBizId') is not None:
             self.merchant_biz_id = m.get('MerchantBizId')
         if m.get('MerchantUserId') is not None:
@@ -2731,6 +2765,246 @@ class InitializeRequest(TeaModel):
             self.show_ocr_result = m.get('ShowOcrResult')
         if m.get('StyleConfig') is not None:
             self.style_config = m.get('StyleConfig')
+        if m.get('UseNFC') is not None:
+            self.use_nfc = m.get('UseNFC')
+        return self
+
+
+class InitializeShrinkRequest(TeaModel):
+    def __init__(
+        self,
+        app_quality_check: str = None,
+        authorize: str = None,
+        callback_token: str = None,
+        callback_url: str = None,
+        crop: str = None,
+        date_of_birth: str = None,
+        date_of_expiry: str = None,
+        doc_page_config_shrink: str = None,
+        doc_scan_mode: str = None,
+        doc_type: str = None,
+        doc_video: str = None,
+        document_number: str = None,
+        experience_code: str = None,
+        face_picture_base_64: str = None,
+        face_picture_url: str = None,
+        id_face_quality: str = None,
+        id_spoof: str = None,
+        id_threshold: str = None,
+        language_config: str = None,
+        mrtdinput: str = None,
+        merchant_biz_id: str = None,
+        merchant_user_id: str = None,
+        meta_info: str = None,
+        model: str = None,
+        ocr: str = None,
+        procedure_priority: str = None,
+        product_code: str = None,
+        product_flow: str = None,
+        return_url: str = None,
+        scene_code: str = None,
+        security_level: str = None,
+        show_album_icon: str = None,
+        show_guide_page: str = None,
+        show_ocr_result: str = None,
+        style_config: str = None,
+        use_nfc: str = None,
+    ):
+        self.app_quality_check = app_quality_check
+        self.authorize = authorize
+        self.callback_token = callback_token
+        self.callback_url = callback_url
+        self.crop = crop
+        self.date_of_birth = date_of_birth
+        self.date_of_expiry = date_of_expiry
+        self.doc_page_config_shrink = doc_page_config_shrink
+        self.doc_scan_mode = doc_scan_mode
+        self.doc_type = doc_type
+        self.doc_video = doc_video
+        self.document_number = document_number
+        self.experience_code = experience_code
+        self.face_picture_base_64 = face_picture_base_64
+        self.face_picture_url = face_picture_url
+        self.id_face_quality = id_face_quality
+        self.id_spoof = id_spoof
+        self.id_threshold = id_threshold
+        self.language_config = language_config
+        self.mrtdinput = mrtdinput
+        self.merchant_biz_id = merchant_biz_id
+        self.merchant_user_id = merchant_user_id
+        self.meta_info = meta_info
+        self.model = model
+        # OCR。
+        self.ocr = ocr
+        self.procedure_priority = procedure_priority
+        self.product_code = product_code
+        self.product_flow = product_flow
+        self.return_url = return_url
+        self.scene_code = scene_code
+        self.security_level = security_level
+        self.show_album_icon = show_album_icon
+        self.show_guide_page = show_guide_page
+        self.show_ocr_result = show_ocr_result
+        self.style_config = style_config
+        self.use_nfc = use_nfc
+
+    def validate(self):
+        pass
+
+    def to_map(self):
+        _map = super().to_map()
+        if _map is not None:
+            return _map
+
+        result = dict()
+        if self.app_quality_check is not None:
+            result['AppQualityCheck'] = self.app_quality_check
+        if self.authorize is not None:
+            result['Authorize'] = self.authorize
+        if self.callback_token is not None:
+            result['CallbackToken'] = self.callback_token
+        if self.callback_url is not None:
+            result['CallbackUrl'] = self.callback_url
+        if self.crop is not None:
+            result['Crop'] = self.crop
+        if self.date_of_birth is not None:
+            result['DateOfBirth'] = self.date_of_birth
+        if self.date_of_expiry is not None:
+            result['DateOfExpiry'] = self.date_of_expiry
+        if self.doc_page_config_shrink is not None:
+            result['DocPageConfig'] = self.doc_page_config_shrink
+        if self.doc_scan_mode is not None:
+            result['DocScanMode'] = self.doc_scan_mode
+        if self.doc_type is not None:
+            result['DocType'] = self.doc_type
+        if self.doc_video is not None:
+            result['DocVideo'] = self.doc_video
+        if self.document_number is not None:
+            result['DocumentNumber'] = self.document_number
+        if self.experience_code is not None:
+            result['ExperienceCode'] = self.experience_code
+        if self.face_picture_base_64 is not None:
+            result['FacePictureBase64'] = self.face_picture_base_64
+        if self.face_picture_url is not None:
+            result['FacePictureUrl'] = self.face_picture_url
+        if self.id_face_quality is not None:
+            result['IdFaceQuality'] = self.id_face_quality
+        if self.id_spoof is not None:
+            result['IdSpoof'] = self.id_spoof
+        if self.id_threshold is not None:
+            result['IdThreshold'] = self.id_threshold
+        if self.language_config is not None:
+            result['LanguageConfig'] = self.language_config
+        if self.mrtdinput is not None:
+            result['MRTDInput'] = self.mrtdinput
+        if self.merchant_biz_id is not None:
+            result['MerchantBizId'] = self.merchant_biz_id
+        if self.merchant_user_id is not None:
+            result['MerchantUserId'] = self.merchant_user_id
+        if self.meta_info is not None:
+            result['MetaInfo'] = self.meta_info
+        if self.model is not None:
+            result['Model'] = self.model
+        if self.ocr is not None:
+            result['Ocr'] = self.ocr
+        if self.procedure_priority is not None:
+            result['ProcedurePriority'] = self.procedure_priority
+        if self.product_code is not None:
+            result['ProductCode'] = self.product_code
+        if self.product_flow is not None:
+            result['ProductFlow'] = self.product_flow
+        if self.return_url is not None:
+            result['ReturnUrl'] = self.return_url
+        if self.scene_code is not None:
+            result['SceneCode'] = self.scene_code
+        if self.security_level is not None:
+            result['SecurityLevel'] = self.security_level
+        if self.show_album_icon is not None:
+            result['ShowAlbumIcon'] = self.show_album_icon
+        if self.show_guide_page is not None:
+            result['ShowGuidePage'] = self.show_guide_page
+        if self.show_ocr_result is not None:
+            result['ShowOcrResult'] = self.show_ocr_result
+        if self.style_config is not None:
+            result['StyleConfig'] = self.style_config
+        if self.use_nfc is not None:
+            result['UseNFC'] = self.use_nfc
+        return result
+
+    def from_map(self, m: dict = None):
+        m = m or dict()
+        if m.get('AppQualityCheck') is not None:
+            self.app_quality_check = m.get('AppQualityCheck')
+        if m.get('Authorize') is not None:
+            self.authorize = m.get('Authorize')
+        if m.get('CallbackToken') is not None:
+            self.callback_token = m.get('CallbackToken')
+        if m.get('CallbackUrl') is not None:
+            self.callback_url = m.get('CallbackUrl')
+        if m.get('Crop') is not None:
+            self.crop = m.get('Crop')
+        if m.get('DateOfBirth') is not None:
+            self.date_of_birth = m.get('DateOfBirth')
+        if m.get('DateOfExpiry') is not None:
+            self.date_of_expiry = m.get('DateOfExpiry')
+        if m.get('DocPageConfig') is not None:
+            self.doc_page_config_shrink = m.get('DocPageConfig')
+        if m.get('DocScanMode') is not None:
+            self.doc_scan_mode = m.get('DocScanMode')
+        if m.get('DocType') is not None:
+            self.doc_type = m.get('DocType')
+        if m.get('DocVideo') is not None:
+            self.doc_video = m.get('DocVideo')
+        if m.get('DocumentNumber') is not None:
+            self.document_number = m.get('DocumentNumber')
+        if m.get('ExperienceCode') is not None:
+            self.experience_code = m.get('ExperienceCode')
+        if m.get('FacePictureBase64') is not None:
+            self.face_picture_base_64 = m.get('FacePictureBase64')
+        if m.get('FacePictureUrl') is not None:
+            self.face_picture_url = m.get('FacePictureUrl')
+        if m.get('IdFaceQuality') is not None:
+            self.id_face_quality = m.get('IdFaceQuality')
+        if m.get('IdSpoof') is not None:
+            self.id_spoof = m.get('IdSpoof')
+        if m.get('IdThreshold') is not None:
+            self.id_threshold = m.get('IdThreshold')
+        if m.get('LanguageConfig') is not None:
+            self.language_config = m.get('LanguageConfig')
+        if m.get('MRTDInput') is not None:
+            self.mrtdinput = m.get('MRTDInput')
+        if m.get('MerchantBizId') is not None:
+            self.merchant_biz_id = m.get('MerchantBizId')
+        if m.get('MerchantUserId') is not None:
+            self.merchant_user_id = m.get('MerchantUserId')
+        if m.get('MetaInfo') is not None:
+            self.meta_info = m.get('MetaInfo')
+        if m.get('Model') is not None:
+            self.model = m.get('Model')
+        if m.get('Ocr') is not None:
+            self.ocr = m.get('Ocr')
+        if m.get('ProcedurePriority') is not None:
+            self.procedure_priority = m.get('ProcedurePriority')
+        if m.get('ProductCode') is not None:
+            self.product_code = m.get('ProductCode')
+        if m.get('ProductFlow') is not None:
+            self.product_flow = m.get('ProductFlow')
+        if m.get('ReturnUrl') is not None:
+            self.return_url = m.get('ReturnUrl')
+        if m.get('SceneCode') is not None:
+            self.scene_code = m.get('SceneCode')
+        if m.get('SecurityLevel') is not None:
+            self.security_level = m.get('SecurityLevel')
+        if m.get('ShowAlbumIcon') is not None:
+            self.show_album_icon = m.get('ShowAlbumIcon')
+        if m.get('ShowGuidePage') is not None:
+            self.show_guide_page = m.get('ShowGuidePage')
+        if m.get('ShowOcrResult') is not None:
+            self.show_ocr_result = m.get('ShowOcrResult')
+        if m.get('StyleConfig') is not None:
+            self.style_config = m.get('StyleConfig')
+        if m.get('UseNFC') is not None:
+            self.use_nfc = m.get('UseNFC')
         return self